# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Restaurant APi",
        "description": "A simple api to manager a Restaurant",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/",
            "description": "Local Dev, or from Heroku"
        },
        {
            "url": "/api/",
            "description": "With docker-compose and nginx proxy"
        }
    ],
    "tags": [
        {
            "name": "Users",
            "description": "API for users in the system"
        },
        {
            "name": "Orders",
            "description": "API for orders in the system"
        },
        {
            "name": "Products",
            "description": "API for products in the system"
        }
    ],
    "paths": {
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id in system",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                },
                                "example": {
                                    "firstName": "Chris",
                                    "lastName": "Dietrich",
                                    "email": "smarterchrisdietrich@gmail.com",
                                    "document": "509-20-8517"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Bad Request Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register user in system",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created User",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "User Created"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Register user in system",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "User Created"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/search": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/FilterOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Find Orders with filter in system",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/OrderSearchResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/OrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Find Orders with filter in system",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/OrderCreationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Find Orders by id in system",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order updated"
                    }
                }
            }
        },
        "/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ProductCreationResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserResponse": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "document": {
                        "type": "string"
                    }
                }
            },
            "UserRequest": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "document": {
                        "type": "string"
                    }
                }
            },
            "OrderRequest": {
                "type": "object",
                "properties": {
                    "user": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "document": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "cellphone": {
                                "type": "string"
                            }
                        }
                    },
                    "deliveryAddress": {
                        "type": "object",
                        "properties": {
                            "city": {
                                "type": "string"
                            },
                            "complement": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            },
                            "neighborhood": {
                                "type": "string"
                            },
                            "number": {
                                "type": "string"
                            },
                            "region": {
                                "type": "string"
                            },
                            "streetAddress": {
                                "type": "string"
                            },
                            "zipCOde": {
                                "type": "string"
                            }
                        }
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "productCode": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "string"
                                },
                                "unitaryValue": {
                                    "type": "string"
                                },
                                "discountValue": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "total": {
                        "type": "string"
                    },
                    "totalProducts": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "shippingValue": {
                        "type": "string"
                    }
                }
            },
            "OrderCreationResponse": {
                "type": "object",
                "properties": {
                    "orderNumber": {
                        "type": "string"
                    },
                    "status": {
                        "type": "number"
                    },
                    "creationDate": {
                        "type": "string"
                    }
                }
            },
            "FilterOrderRequest": {
                "type": "object",
                "properties": {
                    "userDocument": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "productCode": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "streetAddress": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "price": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "OrderSearchResponse": {
                "type": "array",
                "items": {
                    "type": "object",
                    "properties": {
                        "orderNumber": {
                            "type": "string"
                        },
                        "creationDate": {
                            "type": "string"
                        },
                        "user": {
                            "type": "string"
                        },
                        "status": {
                            "type": "string"
                        },
                        "typePayment": {
                            "type": "string"
                        },
                        "total": {
                            "type": "string"
                        }
                    }
                }
            },
            "OrderResponse": {
                "type": "object",
                "properties": {
                    "user": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "document": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "cellphone": {
                                "type": "string"
                            }
                        }
                    },
                    "deliveryAddress": {
                        "type": "object",
                        "properties": {
                            "city": {
                                "type": "string"
                            },
                            "complement": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            },
                            "neighborhood": {
                                "type": "string"
                            },
                            "number": {
                                "type": "string"
                            },
                            "region": {
                                "type": "string"
                            },
                            "streetAddress": {
                                "type": "string"
                            },
                            "zipCOde": {
                                "type": "string"
                            }
                        }
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "productCode": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "string"
                                },
                                "total": {
                                    "type": "string"
                                },
                                "unitaryValue": {
                                    "type": "string"
                                },
                                "discountValue": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "total": {
                        "type": "string"
                    },
                    "totalProducts": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "shippingValue": {
                        "type": "string"
                    },
                    "creationDate": {
                        "type": "string"
                    },
                    "updateDate": {
                        "type": "string"
                    }
                }
            },
            "ProductRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    }
                }
            },
            "ProductCreationResponse": {
                "type": "object",
                "properties": {
                    "productCode": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "dateCreation": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}